week 4 updates

+++ app/templates/filters.py
@@ -3,5 +3,5 @@
 # Formats number values into dollar currency
 # e.g. 123690 is changed to $123,690M
 def format_currency(value):
-	return "${:,}".format(int(round((value),-4)))
+	return "${:,}".format((round((value),-4)))
